requirements.txt
ipykernel
openai>=1.0.0
python-dotenv>=0.19.0
yfinance
pandas
matplotlib
colorama>=0.4.4